Rename layout to better reflect purpose
RISDEV-10287
