+++ frontend/src/pages/advanced-search/index.vue
@@ -214,7 +214,7 @@ watch(searchStatus, async (newStatus, oldStatus) => {
 </script>
 
 <template>
-  <NuxtLayout name="content">
+  <NuxtLayout name="breadcrumb-page">
     <template #breadcrumb>
       <Breadcrumbs :items="[{ label: 'Erweiterte Suche' }]" />
     </template>
