+++ frontend/src/pages/translations/[id].vue
@@ -93,7 +93,7 @@ const detailsTabPanelTitleId = useId();
 </script>
 
 <template>
-  <NuxtLayout name="content">
+  <NuxtLayout name="breadcrumb-page">
     <template #breadcrumb>
       <div class="flex items-center gap-8 print:hidden">
         <Breadcrumbs :items="breadcrumbItems" class="grow" />
